builder/requirements.txt
SCons
Jinja2
pyyaml
numpy
numba
imageio
ConfigArgParse